refactor: Money as abstract class
- Refactore Money interface to abstract class
- Constrainted all its public functions with final to exclude extentding them.
- RUB and USD money implementation extend the abstract Money class
